# src/main/java/com/openbanking/direct/debits/model/OBReadDirectDebit2.java
package com.openbanking.direct.debits.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadDirectDebit2
 */
@Getter
@Setter
public class OBReadDirectDebit2 implements Serializable {
	private static final long serialVersionUID = 1L;

}